requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_iam
aws-cdk.aws_directoryservice
aws-cdk.aws_autoscaling